requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
pypdf
python-docx
python-pptx
pandas
